+++ nas-20170626/alibabacloud_nas20170626/models/_create_lifecycle_policy_request.py
@@ -4,26 +4,31 @@ from __future__ import annotations
 
 from typing import List
 
+from alibabacloud_nas20170626 import models as main_models
 from darabonba.model import DaraModel
 
 class CreateLifecyclePolicyRequest(DaraModel):
     def __init__(
         self,
+        description: str = None,
         file_system_id: str = None,
         lifecycle_policy_name: str = None,
+        lifecycle_policy_type: str = None,
         lifecycle_rule_name: str = None,
         path: str = None,
         paths: List[str] = None,
+        retrieve_rules: List[main_models.CreateLifecyclePolicyRequestRetrieveRules] = None,
         storage_type: str = None,
+        transit_rules: List[main_models.CreateLifecyclePolicyRequestTransitRules] = None,
     ):
+        self.description = description
         # The ID of the file system.
         # 
         # This parameter is required.
         self.file_system_id = file_system_id
         # The name of the lifecycle policy. The name must be 3 to 64 characters in length and can contain letters, digits, underscores (_), and hyphens (-). The name must start with a letter.
-        # 
-        # This parameter is required.
         self.lifecycle_policy_name = lifecycle_policy_name
+        self.lifecycle_policy_type = lifecycle_policy_type
         # The management rule that is associated with the lifecycle policy.
         # 
         # Valid values:
@@ -43,26 +48,41 @@ class CreateLifecyclePolicyRequest(DaraModel):
         # 
         # If you specify this parameter, you can associate the lifecycle policy with multiple directories. Each path must start with a forward slash (/) and must be a path that exists in the mount target. Valid values of N: 1 to 10.
         self.paths = paths
+        self.retrieve_rules = retrieve_rules
         # The storage type of the data that is dumped to the IA storage medium.
         # 
         # Default value: InfrequentAccess (IA).
         # 
         # This parameter is required.
         self.storage_type = storage_type
+        self.transit_rules = transit_rules
 
     def validate(self):
-        pass
+        if self.retrieve_rules:
+            for v1 in self.retrieve_rules:
+                 if v1:
+                    v1.validate()
+        if self.transit_rules:
+            for v1 in self.transit_rules:
+                 if v1:
+                    v1.validate()
 
     def to_map(self):
         result = dict()
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.description is not None:
+            result['Description'] = self.description
+
         if self.file_system_id is not None:
             result['FileSystemId'] = self.file_system_id
 
         if self.lifecycle_policy_name is not None:
             result['LifecyclePolicyName'] = self.lifecycle_policy_name
+
+        if self.lifecycle_policy_type is not None:
+            result['LifecyclePolicyType'] = self.lifecycle_policy_type
 
         if self.lifecycle_rule_name is not None:
             result['LifecycleRuleName'] = self.lifecycle_rule_name
@@ -73,18 +93,34 @@ class CreateLifecyclePolicyRequest(DaraModel):
         if self.paths is not None:
             result['Paths'] = self.paths
 
+        result['RetrieveRules'] = []
+        if self.retrieve_rules is not None:
+            for k1 in self.retrieve_rules:
+                result['RetrieveRules'].append(k1.to_map() if k1 else None)
+
         if self.storage_type is not None:
             result['StorageType'] = self.storage_type
+
+        result['TransitRules'] = []
+        if self.transit_rules is not None:
+            for k1 in self.transit_rules:
+                result['TransitRules'].append(k1.to_map() if k1 else None)
 
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Description') is not None:
+            self.description = m.get('Description')
+
         if m.get('FileSystemId') is not None:
             self.file_system_id = m.get('FileSystemId')
 
         if m.get('LifecyclePolicyName') is not None:
             self.lifecycle_policy_name = m.get('LifecyclePolicyName')
+
+        if m.get('LifecyclePolicyType') is not None:
+            self.lifecycle_policy_type = m.get('LifecyclePolicyType')
 
         if m.get('LifecycleRuleName') is not None:
             self.lifecycle_rule_name = m.get('LifecycleRuleName')
@@ -95,8 +131,90 @@ class CreateLifecyclePolicyRequest(DaraModel):
         if m.get('Paths') is not None:
             self.paths = m.get('Paths')
 
+        self.retrieve_rules = []
+        if m.get('RetrieveRules') is not None:
+            for k1 in m.get('RetrieveRules'):
+                temp_model = main_models.CreateLifecyclePolicyRequestRetrieveRules()
+                self.retrieve_rules.append(temp_model.from_map(k1))
+
         if m.get('StorageType') is not None:
             self.storage_type = m.get('StorageType')
+
+        self.transit_rules = []
+        if m.get('TransitRules') is not None:
+            for k1 in m.get('TransitRules'):
+                temp_model = main_models.CreateLifecyclePolicyRequestTransitRules()
+                self.transit_rules.append(temp_model.from_map(k1))
+
+        return self
+
+class CreateLifecyclePolicyRequestTransitRules(DaraModel):
+    def __init__(
+        self,
+        attribute: str = None,
+        threshold: str = None,
+    ):
+        self.attribute = attribute
+        self.threshold = threshold
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.attribute is not None:
+            result['Attribute'] = self.attribute
+
+        if self.threshold is not None:
+            result['Threshold'] = self.threshold
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Attribute') is not None:
+            self.attribute = m.get('Attribute')
+
+        if m.get('Threshold') is not None:
+            self.threshold = m.get('Threshold')
+
+        return self
+
+class CreateLifecyclePolicyRequestRetrieveRules(DaraModel):
+    def __init__(
+        self,
+        attribute: str = None,
+        threshold: str = None,
+    ):
+        self.attribute = attribute
+        self.threshold = threshold
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.attribute is not None:
+            result['Attribute'] = self.attribute
+
+        if self.threshold is not None:
+            result['Threshold'] = self.threshold
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Attribute') is not None:
+            self.attribute = m.get('Attribute')
+
+        if m.get('Threshold') is not None:
+            self.threshold = m.get('Threshold')
 
         return self