+++ nas-20170626/alibabacloud_nas20170626/models/_describe_lifecycle_policies_request.py
@@ -7,18 +7,25 @@ from darabonba.model import DaraModel
 class DescribeLifecyclePoliciesRequest(DaraModel):
     def __init__(
         self,
+        description: str = None,
         file_system_id: str = None,
+        lifecycle_policy_id: str = None,
         lifecycle_policy_name: str = None,
+        lifecycle_policy_type: str = None,
         page_number: int = None,
         page_size: int = None,
+        path: str = None,
         storage_type: str = None,
     ):
+        self.description = description
         # The ID of the file system.
         self.file_system_id = file_system_id
+        self.lifecycle_policy_id = lifecycle_policy_id
         # The name of the lifecycle policy. The name must meet the following conventions:
         # 
         # The name must be 3 to 64 characters in length and must start with a letter. It can contain letters, digits, underscores (_), and hyphens (-).
         self.lifecycle_policy_name = lifecycle_policy_name
+        self.lifecycle_policy_type = lifecycle_policy_type
         # The page number.
         # 
         # Pages start from page 1. Default value: 1.
@@ -29,6 +36,7 @@ class DescribeLifecyclePoliciesRequest(DaraModel):
         # 
         # Default value: 10.
         self.page_size = page_size
+        self.path = path
         # The storage class.
         # 
         # *   InfrequentAccess: the Infrequent Access (IA) storage class.
@@ -45,17 +53,29 @@ class DescribeLifecyclePoliciesRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.description is not None:
+            result['Description'] = self.description
+
         if self.file_system_id is not None:
             result['FileSystemId'] = self.file_system_id
 
+        if self.lifecycle_policy_id is not None:
+            result['LifecyclePolicyId'] = self.lifecycle_policy_id
+
         if self.lifecycle_policy_name is not None:
             result['LifecyclePolicyName'] = self.lifecycle_policy_name
+
+        if self.lifecycle_policy_type is not None:
+            result['LifecyclePolicyType'] = self.lifecycle_policy_type
 
         if self.page_number is not None:
             result['PageNumber'] = self.page_number
 
         if self.page_size is not None:
             result['PageSize'] = self.page_size
+
+        if self.path is not None:
+            result['Path'] = self.path
 
         if self.storage_type is not None:
             result['StorageType'] = self.storage_type
@@ -64,17 +84,29 @@ class DescribeLifecyclePoliciesRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Description') is not None:
+            self.description = m.get('Description')
+
         if m.get('FileSystemId') is not None:
             self.file_system_id = m.get('FileSystemId')
 
+        if m.get('LifecyclePolicyId') is not None:
+            self.lifecycle_policy_id = m.get('LifecyclePolicyId')
+
         if m.get('LifecyclePolicyName') is not None:
             self.lifecycle_policy_name = m.get('LifecyclePolicyName')
+
+        if m.get('LifecyclePolicyType') is not None:
+            self.lifecycle_policy_type = m.get('LifecyclePolicyType')
 
         if m.get('PageNumber') is not None:
             self.page_number = m.get('PageNumber')
 
         if m.get('PageSize') is not None:
             self.page_size = m.get('PageSize')
+
+        if m.get('Path') is not None:
+            self.path = m.get('Path')
 
         if m.get('StorageType') is not None:
             self.storage_type = m.get('StorageType')